+++ dotnet-rest-api/src/LoginEKO.FileProcessingService.Domain/Models/FileDto.cs
@@ -11,7 +11,7 @@ namespace LoginEKO.FileProcessingService.Domain.Models
         public Guid Id { get; set; }
         public required string Filename { get; set; }
         public required string Extension { get; set; }
-        public double SizeInBytes { get; set; }
+        public long SizeInBytes { get; set; }
         public required byte[] BinaryObject { get; set; }
         public string MD5Hash { get; set; }
         public DateTime CreatedDate { get; set; }
